Move to Twirl for template parsing.
Fixed tests that started failing under 2.11.

* added a strategic `initialize` call before checking `.isAbstractClass`, since
  flags weren’t there yet
* made the stub `play.mvc.Result` a proper public class (default visibility in
  Java is package private). The 2.11 compiler correctly failed those tests.
