+++ org.scala-ide.play2.tests/.classpath
@@ -4,5 +4,7 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.6"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="lib" path="target/lib/mockito-all-1.9.0.jar"/>
+	<classpathentry kind="lib" path="target/lib/twirl-compiler.jar"/>
+	<classpathentry kind="lib" path="target/lib/twirl-parser.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
