+++ org.scala-ide.play2.tests/test-workspace/routeActionCompletions/src/play/mvc/Result.java
@@ -1,3 +1,3 @@
 package play.mvc;
 
-abstract class Result {}
+abstract public class Result {}
