+++ org.scala-ide.play2/META-INF/MANIFEST.MF
@@ -43,12 +43,11 @@ Require-Bundle:
  org.eclipse.wst.jsdt.core;bundle-version="1.1.102"
 Import-Package: 
  com.ibm.icu.text;apply-aspects:=false;org.eclipse.swt.graphics;apply-aspects:=false,
- scala.tools.eclipse.contribution.weaving.jdt.ui.javaeditor.formatter;apply-aspects:=false
+ scala.tools.eclipse.contribution.weaving.jdt.ui.javaeditor.formatter;apply-aspects:=false,
+ scala.xml,
+ scala.util.parsing.input
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-Activator: org.scalaide.play2.PlayPlugin
 Bundle-ClassPath: .,
- target/lib/play.jar,
- target/lib/scala-arm.jar,
- target/lib/scala-io-core.jar,
- target/lib/scala-io-file.jar,
- target/lib/templates.jar
+ target/lib/twirl-compiler.jar,
+ target/lib/twirl-parser.jar
