+++ org.scala-ide.play2/build.properties
@@ -7,4 +7,6 @@ bin.includes = META-INF/,\
                icons/,\
                about.ini,\
                about.png,\
-               target/lib/
+               target/lib/,\
+               target/lib/twirl-compiler.jar,\
+               target/lib/twirl-parser.jar
