+++ org.scala-ide.play2/src/org/scalaide/play2/templateeditor/compiler/CompilerUsing.scala
@@ -4,18 +4,18 @@ import java.io.File
 import scala.util.Failure
 import scala.util.Try
 import org.scalaide.play2.PlayProject
-import play.templates.GeneratedSourceVirtual
-import play.templates.ScalaTemplateCompiler
-import play.templates.ScalaTemplateCompiler._
-import play.templates.TemplateCompilationError
-import scalax.file.Path
+import play.twirl.compiler.GeneratedSourceVirtual
+import play.twirl.compiler.TwirlCompiler
+import play.twirl.compiler.TwirlCompiler._
+import play.twirl.compiler.TemplateCompilationError
 import org.scalaide.play2.properties.PlayPreferences
 import org.scalaide.logging.HasLogger
+import scala.io.Codec
 /**
  * a helper for using template compiler
  */
 object CompilerUsing extends HasLogger {
-  val templateCompiler = ScalaTemplateCompiler
+  val templateCompiler = TwirlCompiler
   val defaultTemplateImports = """
 import models._
 import controllers._
@@ -44,6 +44,7 @@ import views.html._
         "play.api.templates.Html",
         "play.api.templates.HtmlFormat",
         defaultTemplateImports + playProject.additionalTemplateImports(extension),
+        Codec.default,
         inclusiveDot
       )
     } recoverWith {
@@ -83,7 +84,7 @@ case class TemplateToScalaCompilationError(source: File, message: String, offset
 
 object PositionHelper {
   def convertLineColumnToOffset(source: File, line: Int, column: Int): Int = {
-    convertLineColumnToOffset(Path(source).string, line, column)
+    convertLineColumnToOffset(scala.io.Source.fromFile(source).mkString, line, column)
   }
 
   def convertLineColumnToOffset(content: String, line: Int, column: Int): Int = {
