+++ org.scala-ide.play2/src/org/scalaide/play2/templateeditor/compiler/TemplatePresentationCompiler.scala
@@ -1,29 +1,24 @@
 package org.scalaide.play2.templateeditor.compiler
 
 import org.eclipse.jdt.core.compiler.IProblem
-import org.scalaide.play2.PlayProject
 import org.scalaide.play2.templateeditor.TemplateCompilationUnit
-//import scala.tools.eclipse.javaelements.ScalaSourceFile
 import org.eclipse.jdt.internal.compiler.problem.ProblemSeverities
 import org.eclipse.jdt.internal.compiler.problem.DefaultProblem
-//import scala.tools.eclipse.javaelements.ScalaCompilationUnit
-import scalax.file.Path
 import scala.tools.nsc.util.SourceFile
 import scala.tools.nsc.util.BatchSourceFile
-import play.templates.GeneratedSource
+import play.twirl.compiler.GeneratedSource
 import scala.tools.nsc.io.AbstractFile
 import scala.tools.nsc.io.VirtualFile
 import scala.tools.nsc.io.PlainFile
 import java.io.File
-//import scala.tools.eclipse.util.EclipseFile
-//import scala.tools.eclipse.util.EclipseResource
 import org.scalaide.core.compiler.ScalaPresentationCompiler
-import play.templates.GeneratedSourceVirtual
+import play.twirl.compiler.GeneratedSourceVirtual
 import org.scalaide.logging.HasLogger
 import scala.util.Failure
 import scala.util.Success
 import scala.util.Try
 import scala.collection.mutable
+import org.scalaide.play2.PlayProject
 
 /**
  * presentation compiler for template files
@@ -118,7 +113,7 @@ class TemplatePresentationCompiler(playProject: PlayProject) extends HasLogger {
     }).flatten
 
   def destroy() = {
-    CompilerUsing.templateCompiler.TemplateAsFunctionCompiler.shutdownPresentationCompiler()
+    CompilerUsing.templateCompiler.TemplateAsFunctionCompiler.PresentationCompiler.shutdown()
   }
 }
