+++ pom.xml
@@ -9,7 +9,7 @@
   <parent>
     <groupId>org.scala-ide</groupId>
     <artifactId>plugin-profiles</artifactId>
-    <version>1.0.2</version>
+    <version>1.0.5</version>
   </parent>
 
   <groupId>org.scala-ide</groupId>
@@ -29,55 +29,24 @@
   </modules>
 
   <properties>
+    <scala.binary.version>${scala.version.short}</scala.binary.version>
     <!-- play version depends on Scala version. 2.10.x -> 2.1.x !-->
-    <play.artifactId>play_2.10</play.artifactId>
-    <play.version>2.1.1</play.version>
-    <templates.artifactId>templates_2.10</templates.artifactId>
-    <templates.version>2.1.1</templates.version>
-    <scala.io.core.artifactId>scala-io-core_2.10</scala.io.core.artifactId>
-    <scala.io.core.version>0.4.2</scala.io.core.version>
-    <scala.io.file.artifactId>scala-io-file_2.10</scala.io.file.artifactId>
-    <scala.io.file.version>0.4.2</scala.io.file.version>
-    <scala.arm.groupId>com.jsuereth</scala.arm.groupId>
-    <scala.arm.artifactId>scala-arm_2.10</scala.arm.artifactId>
-    <scala.arm.version>1.3</scala.arm.version>
+    <twirl-compiler.artifactId>twirl-compiler_${scala.binary.version}</twirl-compiler.artifactId>
+    <twirl-parser.artifactId>twirl-parser_${scala.binary.version}</twirl-parser.artifactId>
+    <twirl.version>1.0.2</twirl.version>
   </properties>
-
-  <repositories>
-    <repository>
-      <id>play.release</id>
-      <name>play releases and snapshots</name>
-      <url>http://repo.typesafe.com/typesafe/releases/</url>
-      <snapshots><enabled>false</enabled></snapshots>
-    </repository>
-  </repositories>
 
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>play</groupId>
-        <artifactId>${play.artifactId}</artifactId>
-        <version>${play.version}</version>
+        <groupId>com.typesafe.play</groupId>
+        <artifactId>${twirl-compiler.artifactId}</artifactId>
+        <version>${twirl.version}</version>
       </dependency>
       <dependency>
-        <groupId>play</groupId>
-        <artifactId>${templates.artifactId}</artifactId>
-        <version>${templates.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.scala-incubator.io</groupId>
-        <artifactId>${scala.io.core.artifactId}</artifactId>
-        <version>${scala.io.core.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.scala-incubator.io</groupId>
-        <artifactId>${scala.io.file.artifactId}</artifactId>
-        <version>${scala.io.file.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>${scala.arm.groupId}</groupId>
-        <artifactId>${scala.arm.artifactId}</artifactId>
-        <version>${scala.arm.version}</version>
+        <groupId>com.typesafe.play</groupId>
+        <artifactId>${twirl-parser.artifactId}</artifactId>
+        <version>${twirl.version}</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
